+++ src/main/java/com/fractals/beans/Track.java
@@ -98,17 +98,17 @@ public class Track implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "cost_price")
-    private float costPrice;
+    private double costPrice;
     
     @Basic(optional = false)
     @NotNull
     @Column(name = "list_price")
-    private float listPrice;
+    private double listPrice;
     
     @Basic(optional = false)
     @NotNull
     @Column(name = "sale_price")
-    private float salePrice;
+    private double salePrice;
     
     @Column(name = "removed_at")
     @Temporal(TemporalType.TIMESTAMP)
@@ -150,7 +150,7 @@ public class Track implements Serializable {
         this.id = id;
     }
 
-    public Track(Integer id, String title, String songwriter, String duration, int albumNum, String coverFile, Date createdAt, float costPrice, float listPrice, float salePrice, boolean available, boolean isIndividual) {
+    public Track(Integer id, String title, String songwriter, String duration, int albumNum, String coverFile, Date createdAt, double costPrice, double listPrice, double salePrice, boolean available, boolean isIndividual) {
         this.id = id;
         this.title = title;
         this.songwriter = songwriter;
@@ -221,27 +221,27 @@ public class Track implements Serializable {
         this.createdAt = createdAt;
     }
 
-    public float getCostPrice() {
+    public double getCostPrice() {
         return costPrice;
     }
 
-    public void setCostPrice(float costPrice) {
+    public void setCostPrice(double costPrice) {
         this.costPrice = costPrice;
     }
 
-    public float getListPrice() {
+    public double getListPrice() {
         return listPrice;
     }
 
-    public void setListPrice(float listPrice) {
+    public void setListPrice(double listPrice) {
         this.listPrice = listPrice;
     }
 
-    public float getSalePrice() {
+    public double getSalePrice() {
         return salePrice;
     }
 
-    public void setSalePrice(float salePrice) {
+    public void setSalePrice(double salePrice) {
         this.salePrice = salePrice;
     }
